requirements.txt
pandas
openpyxl
streamlit
langchain
boto3
faiss-cpu
awscli
python-dotenv